requirements.txt
pyinquirer==1.0.3
tinydb==3.15.2
pycryptodome==3.9.7
